+++ packages/local-prover/index.ts
@@ -1,1 +1,2 @@
 export * from "./src/joinsplit";
+export * from "./src/subtreeUpdate";